examples/svgtiny_display_x11.c: include the system copy of svgtiny.h
The header of this file includes instructions for how to build it:

  Compile using:
    gcc -g -W -Wall -o svgtiny_display_x11 svgtiny_display_x11.c \
            `pkg-config --cflags --libs libsvgtiny cairo` -lX11

That pkg-config command will generate the flags to link against the
installed copy of libsvgtiny. The line,

  #include "svgtiny.h"

on the other hand, attempts to use a local header. This commit changes
that line to,

  #include <svgtiny.h>

which will use the corresponding system header from whatever include
directory pkg-config hands us for libsvgtiny.